Фикс (#5307)

+++ code/modules/mob/mob.dm
@@ -1058,17 +1058,19 @@ note dizziness decrements automatically in the mob's Life() proc.
 
 /mob/proc/RemoveSpell(obj/effect/proc_holder/spell/S)
 	spell_list -= S
-	qdel(S)
-	mind.spell_list -= S
+	if(mind)
+		mind.spell_list -= S
 	qdel(S)
 
 /mob/proc/ClearSpells()
 	for(var/spell in spell_list)
 		spell_list -= spell
 		qdel(spell)
-	for(var/spell in mind.spell_list)
-		mind.spell_list -= spell
-		qdel(spell)
+
+	if(mind)
+		for(var/spell in mind.spell_list)
+			mind.spell_list -= spell
+			qdel(spell)
 
 /mob/proc/set_EyesVision(preset = null, transition_time = 5)
 	if(!client) return
